网贷天眼

+++ src/main/java/com/apd/www/service/ProjectService.java
@@ -23,7 +23,7 @@ public interface ProjectService {
 
     public List<Project> getWDTYProjectList(Integer status,String time_from,String time_to,Integer page_size,Integer page_index) throws ParseException;
 
-    public  Integer getWDTYProjectListCount(Integer status,String time_from,String time_to) throws ParseException;
+    public  Long getWDTYProjectListCount(Integer status,String time_from,String time_to) throws ParseException;
 
 
     public List<Project> getYong360ProjectList(Integer page_size,Integer page_index) throws ParseException;
